requirements.txt
fastapi
uvicorn
pydantic
torch
sentence-transformers
transformers
pinecone-client